PowerSimulationsDynamics

API documentation

Index

Exported

PowerSimulationsDynamics.Simulation — Type

Initializes the simulations and builds the indexing. The input system is not modified during the initialization

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

system_to_file::Bool: Serializes the original input system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debugg

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.Simulation! — Function

Builds the simulation object and conducts the indexing process. The initial conditions are stored in the system.

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

system_to_file::Bool: Serializes the initialized system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.get_initial_conditions — Method

Returns a Dictionary with the resulting initial conditions of the simulation

PowerSimulationsDynamics.get_state_series — Method

Function to obtain series of states out of DAE Solution. It receives the solution, the dynamical system and a tuple containing the symbol name of the Dynamic Injection device and the symbol of the state.

PowerSimulationsDynamics.get_voltagemag_series — Method

Function to obtain the voltage magnitude series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

PowerSimulationsDynamics.print_device_states — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.

Internal

PowerSimulationsDynamics.BUILD_STATUS — Type

Defines the status of the simulation object

PowerSimulationsDynamics.generator_inner_vars — Type

Generator Inner Vars:

τe_var :: Electric torque

τm_var :: Mechanical torque

Vf_var :: Field voltage

Vpssvar :: Additional PSS voltage

VRgenvar :: Real part of the terminal voltage

VIgenvar :: Imaginary part of the terminal voltage

ψd_var :: Stator Flux (if defined) in the d-axis

ψq_var :: Stator Flux (if defined) in the q-axis

PowerSimulationsDynamics.inverter_inner_vars — Type

Inverter Inner Vars: @enum inverterinnervars begin mdvar :: Modulation signal on the d-component mqvar :: Modulation signal on the q-component Vdcvar :: DC voltage supplied by the DC source Vrfiltervar :: Voltage seen in the capacitor of the filter in the R-component Vifiltervar :: Voltage seen in the capacitor of the filter in the I-component ωfreqestimatorvar :: Frequency estimated by the frequency estimator. Vocvar :: Control voltage supplied from the outer loop control to the inner loop ωocvar :: Control frequency supplied from the outer loop control the inner loop θocvar :: Variation of the angle (PLL or VSM) of the inverter VRinvvar :: Real terminal voltage on the inverter VIinvvar :: Imaginary terminal voltage on the inverter Vrcnvvar :: Voltage supplied from the converter in the R-component Vicnvvar :: Voltage supplied from the converter in the I-component

PowerSimulationsDynamics.configure_logging — Method

configure_logging(;
    console_level = Logging.Error,
    file_level = Logging.Info,
    filename = "power-simulations.log",
)

Creates console and file loggers.

Note: Log messages may not be written to the file until flush() or close() is called on the returned logger.

Arguments

console_level = Logging.Error: level for console messages

file_level = Logging.Info: level for file messages

filename::String = power-simulations.log: log file

Example

logger = configure_logging(console_level = Logging.Info)
@info "log message"
close(logger)

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 0-state synchronous (classic model) machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.kirchoff_laws! — Method

Kirchoff law for buses.
I_gen_r[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.
I_gen_i[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.

PowerSimulationsDynamics.make_device_index! — Method

Indexes the devices states and maps to the ports

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 6-state (AndersonFouadMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 0-state synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 6-state (MarconatoMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 4-state (SimpleAFMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 4-state (SimpleMarconatoMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.saturation_function — Method

Saturation function for exponential saturation models for machines
    Se(x) = B * x^A

PowerSimulationsDynamics.saturation_function — Method

Saturation function for quadratic saturation models for machines
    Se(x) = B * (x - A)^2 / x