Internal

PowerSimulationsDynamics.BUILD_STATUS — Type

Defines the status of the simulation object

PowerSimulationsDynamics.BranchWrapper — Type

Wraps DynamicBranch devices from PowerSystems to handle changes in controls and connection status, and allocate the required indexes of the state space.

PowerSimulationsDynamics.DynamicWrapper — Type

Wraps DynamicInjection devices from PowerSystems to handle changes in controls and connection status, and allocate the required indexes of the state space.

PowerSimulationsDynamics.SimulationInputs — Method

SimulationInputs build function for MassMatrixModels

PowerSimulationsDynamics.SimulationInputs — Method

SimulationInputs build function for ResidualModels

PowerSimulationsDynamics.generator_inner_vars — Type

Generator Inner Vars:

τe_var :: Electric torque

τm_var :: Mechanical torque

Vf_var :: Field voltage

V_pss_var :: Additional PSS voltage

VR_gen_var :: Real part of the terminal voltage

VI_gen_var :: Imaginary part of the terminal voltage

ψd_var :: Stator Flux (if defined) in the d-axis

ψq_var :: Stator Flux (if defined) in the q-axis

PowerSimulationsDynamics.inverter_inner_vars — Type

Inverter Inner Vars:

md_var :: Modulation signal on the d-component

mq_var :: Modulation signal on the q-component

Vdc_var :: DC voltage supplied by the DC source

Vr_filter_var :: Voltage seen in the capacitor of the filter in the R-component

Vi_filter_var :: Voltage seen in the capacitor of the filter in the I-component

θ_freq_estimator_var :: Angle estimated by the frequency estimator.

ω_freq_estimator_var :: Frequency estimated by the frequency estimator.

V_oc_var :: Control voltage reference in the d-axis supplied from the outer loop control to the inner loop (for Voltage Mode Control)

Id_oc_var :: Control current reference in the d-axis supplied from the outer loop control to the inner loop (for Current Mode Control)

Iq_oc_var :: Control current reference in the q-axis supplied from the outer loop control to the inner loop (for Current Mode Control)

Id_ic_var :: Control current reference in the d-axis supplied from the inner loop control to the converter (for Generic Models)

Iq_ic_var :: Control current reference in the q-axis supplied from the inner loop control to the converter (for Generic Models)

Ir_cnv_var :: Control current reference in the R-axis supplied from the converter to the filter (for Generic Models)

Ii_cnv_var :: Control current reference in the I-axis supplied from the converter to the filter (for Generic Models)

ω_oc_var :: Control frequency supplied from the outer loop control the inner loop

θ_oc_var :: Variation of the angle (PLL or VSM) of the inverter

Vr_inv_var :: Real terminal voltage on the inverter

Vi_inv_var :: Imaginary terminal voltage on the inverter

Vr_cnv_var :: Voltage supplied from the converter in the R-component

Vi_cnv_var :: Voltage supplied from the converter in the I-component

P_ES_var :: Power supplied from the Energy Source side

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a Classic Machine model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a One-D-One-Q model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a Marconato or AndersonFouad model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a GENROU/GENROE model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a GENSAL/GENSAE model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._machine_current — Method

Function to obtain the output current time series of a SimpleMarconato or SimpleAndersonFouad model out of the DAE Solution. It is dispatched via the machine type.

PowerSimulationsDynamics._make_reduced_jacobian_index — Method

Finds the location of the differential states in the reduced Jacobian

PowerSimulationsDynamics._output_current — Method

Function to obtain the output current time series of a LCL Filter model out of the DAE Solution. It is dispatched via the Filter type.

PowerSimulationsDynamics._output_current — Method

Function to obtain the output current time series of a REGCA converter model out of the DAE Solution. It is dispatched via the Converter type.

PowerSimulationsDynamics._post_proc_state_series — Method

Internal function to obtain as a Vector of Float64 of a specific state. It receives the solution and the global index for a state.

PowerSimulationsDynamics.compute_output_current — Method

Function to obtain the output current time series of a Dynamic Generator model out of the DAE Solution. It receives the simulation inputs, the dynamic device and bus voltage. It is dispatched for device type to compute the specific current.

PowerSimulationsDynamics.compute_output_current — Method

Function to obtain the output current time series of a Dynamic Inverter model out of the DAE Solution. It receives the simulation inputs, the dynamic device and bus voltage. It is dispatched for device type to compute the specific current.

PowerSimulationsDynamics.configure_logging — Method

configure_logging(;
    console_level = Logging.Error,
    file_level = Logging.Info,
    filename = "power-simulations.log",
)

Creates console and file loggers.

Note: Log messages may not be written to the file until flush() or close() is called on the returned logger.

Arguments

console_level = Logging.Error: level for console messages

file_level = Logging.Info: level for file messages

filename::String = power-simulations.log: log file

Example

logger = configure_logging(console_level = Logging.Info)
@info "log message"
close(logger)

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 0-state synchronous (classic model) machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mass_matrix_entries — Method

Default implementation of mass matrix entries. Keeps the default values in the identity matrix

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 6-state (AndersonFouadMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 0-state synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 6-state (MarconatoMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 4-state (SimpleAFMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.mdl_machine_ode! — Method

Model of 4-state (SimpleMarconatoMachine) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.post_proc_activepower_series — Method

Function to compute the active power output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.post_proc_imaginary_current_series — Method

Function to compute the imaginary current output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.post_proc_reactivepower_series — Method

Function to compute the active power output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.post_proc_real_current_series — Method

Function to compute the real current output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.post_proc_state_series — Method

Function to obtain the state time series of a specific state. It receives the simulation, and a tuple containing the name of the Dynamic Device and the symbol of the state.

PowerSimulationsDynamics.post_proc_voltage_current_series — Method

Function to obtain voltage and output currents for a dynamic device. It receives the simulation, and the name of the Dynamic Device.

PowerSimulationsDynamics.post_proc_voltage_series — Method

Function to obtain voltage using the bus index (and not the bus number). It receives the solution, the bus index and the total number of buses.

PowerSimulationsDynamics.saturation_function — Method

Saturation function for exponential saturation models for machines
    Se(x) = B * x^A

PowerSimulationsDynamics.saturation_function — Method

Saturation function for quadratic saturation models for machines
    Se(x) = B * (x - A)^2 / x